refactor: separate bank.go into bank_interface.go and command_center.go
- Split repository/bank.go into two focused files
- bank_interface.go: contains Bank struct and database operations
- command_center.go: contains CommandCenter interface and implementation
- Improves code organization and separation of concerns
