+++ repository/bank_interface.go
@@ -38,22 +38,3 @@ func (b *Bank) UpdateAccessCodes(codes []string, siteID int) error {
 func (b *Bank) NewCommandCenterClient(siteID int, ctx context.Context) CommandCenterClient {
 	return &CommandCenter{}
 }
-
-// CommandCenterClient interface for command center operations
-type CommandCenterClient interface {
-	RevokeAccessCodes(revokeUnits []int, options map[string]struct{}) error
-	SetAccessCodes(units []int, options map[string]struct{}) error
-}
-
-// CommandCenter implements CommandCenterClient
-type CommandCenter struct{}
-
-// RevokeAccessCodes revokes access codes for given units
-func (cc *CommandCenter) RevokeAccessCodes(revokeUnits []int, options map[string]struct{}) error {
-	return nil
-}
-
-// SetAccessCodes sets access codes for given units
-func (cc *CommandCenter) SetAccessCodes(units []int, options map[string]struct{}) error {
-	return nil
-}
